ai-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
numpy
scikit-learn
Pillow
opencv-python-headless
requests
httpx
python-dotenv
celery
redis
joblib
tenacity
python-jose[cryptography]==3.3.0
